+++ silo-core/test/foundry/echidna-findings/EchidnaScenarios.i.sol
@@ -1,45 +1,14 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity ^0.8.0;
 
-import "forge-std/Test.sol";
 import {Strings} from "openzeppelin-contracts/utils/Strings.sol";
-
 import {ISiloConfig} from "silo-core/contracts/interfaces/ISiloConfig.sol";
-import {ISilo} from "silo-core/contracts/interfaces/ISilo.sol";
-import {IShareToken} from "silo-core/contracts/interfaces/IShareToken.sol";
-
-import {MintableToken} from "../_common/MintableToken.sol";
-import {SiloLittleHelper} from "../_common/SiloLittleHelper.sol";
-import {SiloFixture, SiloConfigOverride} from "../_common/fixtures/SiloFixture.sol";
+import {EchidnaMiddleman} from "./EchidnaMiddleman.sol";
 
 /*
     forge test -vv --ffi --mc EchidnaScenariosTest
 */
-contract EchidnaScenariosTest is SiloLittleHelper, Test {
-    uint256 constant ACTORS_COUNT = 3;
-    mapping (uint256 index => address actor) actors;
-
-    ISiloConfig siloConfig;
-
-    constructor() {
-        actors[0] = makeAddr("Index 0");
-        actors[1] = makeAddr("Index 1");
-        actors[2] = makeAddr("Index 2");
-    }
-
-    function setUp() public {
-        siloConfig = _setUpLocalFixture("Echidna_MOCK");
-
-        assertTrue(siloConfig.getConfig(address(silo0)).maxLtv != 0, "we need borrow to be allowed");
-
-        token0.setOnDemand(true);
-        token1.setOnDemand(true);
-
-        // same block and time as for E2E Echidna
-        vm.warp(1706745600);
-        vm.roll(17336000);
-    }
-
+contract EchidnaScenariosTest is EchidnaMiddleman {
     /*
 mint(uint8,bool,uint256): passing
 maxBorrow_correctReturnValue(uint8): failed!💥
@@ -62,53 +31,137 @@ error Revert AboveMaxLtv ()
 error Revert AboveMaxLtv ()
 
     forge test -vv --ffi --mt test_cover_echidna_scenario_1
+
+    this test case covers the bug we had in maxBorrow
     */
     function test_cover_echidna_scenario_1() public {
-        //        depositNeverMintsZeroShares(0,false,39962600816669483677151)
-        vm.prank(_chooseActor(0));
-        silo1.deposit(39962600816669483677151, _chooseActor(0));
-        //        depositNeverMintsZeroShares(161,true,22168924613129761549643809883710869859261573373213864899764932836300336298504)
-        vm.prank(_chooseActor(161));
-        silo0.deposit(22168924613129761549643809883710869859261573373213864899764932836300336298504, _chooseActor(161));
-        //        borrow(2,false,39858486983777211695540)
-        vm.prank(_chooseActor(2));
-        silo1.borrow(39858486983777211695540, _chooseActor(2), _chooseActor(2));
+        __depositNeverMintsZeroShares(0,false,39962600816669483677151);
+        __depositNeverMintsZeroShares(161,true,22168924613129761549643809883710869859261573373213864899764932836300336298504);
+        __borrow(2,false,39858486983777211695540);
+
         //        *wait* Time delay: 1 seconds Block delay: 1
         vm.warp(block.timestamp + 1);
-        //        maxBorrow_correctReturnValue(0)
 
-        address actor = _chooseActor(0);
-        uint256 maxAssets = silo0.maxBorrow(actor);
+        (uint256 maxAssets,) = __maxBorrow_correctReturnValue(0);
 
-        (address protected, address collateral, ) = siloConfig.getShareTokens(address(silo1));
-        emit log_named_decimal_uint("collateral shares", IShareToken(collateral).balanceOf(actor), 18);
-        emit log_named_decimal_uint("maxAssets", maxAssets, 18);
-        // assertEq(maxAssets, 339682116, "maxAssets from echidna simulation"); // TODO why echidna shows 339682116?
-
-        vm.prank(actor);
-        silo0.borrow(maxAssets, actor, actor); // should not revert!
+        assertEq(maxAssets, 33968211591454069532969, "echidna borrow amount (339682116)");
     }
 
-    function _chooseActor(
-        uint256 value
-    ) internal returns (address) {
-        uint256 low = 0;
-        uint256 high = ACTORS_COUNT - 1;
+    /*
+maxRedeem_correctMax(uint8): failed!💥
+  Call sequence, shrinking 316/500:
+    mint(220,true,39843190113244004124792096928447682291)
+    depositNeverMintsZeroShares(2,false,22786647965505400975764605899155696411742225287612091359740640699919114746382)
+    borrowShares(32,true,20769187434139310514121985316880385)
+    debtSharesNeverLargerThanDebt() Time delay: 2369 seconds Block delay: 4358
+    maxRedeem_correctMax(4)
 
-        if (value < low || value > high) {
-            uint ans = low + (value % (high - low + 1));
-            string memory valueStr = Strings.toString(value);
-            string memory ansStr = Strings.toString(ans);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                valueStr,
-                " to ",
-                ansStr
-            );
-            emit log(string(message));
-            return actors[ans];
-        }
+Revert NotEnoughLiquidity
 
-        return actors[value];
+    forge test -vv --ffi --mt test_cover_echidna_scenario_2
+
+    this case covers the bug with maxRedeem
+    */
+    function test_cover_echidna_scenario_2() public {
+        __mint(220,true,39843190113244004124792096928447682291);
+        __depositNeverMintsZeroShares(2,false,22786647965505400975764605899155696411742225287612091359740640699919114746382);
+        __borrowShares(32,true,20769187434139310514121985316880385);
+        // Time delay: 2369 seconds Block delay: 4358
+        __timeDelay(2369, 4358);
+
+        __debtSharesNeverLargerThanDebt();
+
+        __maxRedeem_correctMax(4);
+    }
+
+    /*
+maxBorrowShares_correctReturnValue(uint8): failed!💥
+Call sequence, shrinking 294/500:
+deposit(1,false,18362350053917916671701463106358)
+previewDeposit_doesNotReturnMoreThanDeposit(44,5021491078257170633099496333274079409660413512867881700899062515742671857130)
+mint(0,false,127287174252953083636439506782)
+maxBorrow_correctReturnValue(21)
+maxWithdraw_correctMax(110)
+previewDeposit_doesNotReturnMoreThanDeposit(2,62718039246364723308705975150828086672) Time delay: 3865 seconds Block delay: 1791
+maxBorrowShares_correctReturnValue(1)
+
+Revert AboveMaxLtv ()
+
+    forge test -vv --ffi --mt test_cover_echidna_scenario_3
+
+    this case covers the bug with maxBorrowShare
+    */
+    function test_cover_echidna_scenario_3() public {
+        __deposit(1,false,18362350053917916671701463106358);
+        __previewDeposit_doesNotReturnMoreThanDeposit(44,5021491078257170633099496333274079409660413512867881700899062515742671857130);
+        __mint(0,false,127287174252953083636439506782);
+        __maxBorrow_correctReturnValue(21);
+        __maxWithdraw_correctMax(110);
+
+        __timeDelay(3865, 1791);
+
+        __previewDeposit_doesNotReturnMoreThanDeposit(2,62718039246364723308705975150828086672); // Time delay: 3865 seconds Block delay: 1791
+
+        __maxBorrowShares_correctReturnValue(1);
+    }
+
+    /*
+maxRedeem_correctMax(uint8): failed!💥
+  Call sequence:
+    depositNeverMintsZeroShares(161,true,22168924613129761549643809883710869859261573373213864899764932836300336298504)
+    depositNeverMintsZeroShares(0,false,22033284938236427192887)
+    borrow(2,false,143)
+    maxRedeem_correctMax(0)
+    deposit(0,false,8889521246853316927242844750202102194149874260909366919319008) Time delay: 76512 seconds Block delay: 1269
+    maxRedeem_correctMax(0)
+
+
+    forge test -vv --ffi --mt test_cover_echidna_scenario_4
+
+    this test covers maxRedeem bug
+    */
+    function test_cover_echidna_scenario_4() public {
+        __depositNeverMintsZeroShares(161,true,22168924613129761549643809883710869859261573373213864899764932836300336298504);
+        __depositNeverMintsZeroShares(0,false,22033284938236427192887);
+        __borrow(2,false,143);
+        __maxRedeem_correctMax(0);
+
+        __timeDelay(76512, 1269);
+        __deposit(0,false,8889521246853316927242844750202102194149874260909366919319008);
+
+        __maxRedeem_correctMax(0);
+    }
+
+    /*
+maxWithdraw_correctMax(uint8): failed!💥
+  Call sequence, shrinking 318/500:
+    depositNeverMintsZeroShares(161,true,22168924613129761549643809883710869859261573373213864899764932836300336298504)
+    depositNeverMintsZeroShares(0,false,4584498224059781313847754095738)
+    borrow(2,false,382)
+    maxRedeem_correctMax(0)
+    deposit(1,false,4)
+    accrueInterest(false) Time delay: 159643 seconds Block delay: 16746
+    depositNeverMintsZeroShares(1,false,37148972519749285132903713312934615554658075164406276880551013092746735413)
+    maxBorrow_correctReturnValue(0)
+    maxWithdraw_correctMax(0)
+
+
+    forge test -vv --ffi --mt test_cover_echidna_scenario_5
+
+    this test covers bug with maxWithdraw
+    */
+    function test_cover_echidna_scenario_5() public {
+        __depositNeverMintsZeroShares(161,true,22168924613129761549643809883710869859261573373213864899764932836300336298504);
+        __depositNeverMintsZeroShares(0,false,4584498224059781313847754095738);
+        __borrow(2,false,382);
+        __maxRedeem_correctMax(0);
+        __deposit(1,false,4);
+
+        __timeDelay(159643, 16746);
+
+        __accrueInterest(false); // Time delay: 159643 seconds Block delay: 16746
+        __depositNeverMintsZeroShares(1,false,37148972519749285132903713312934615554658075164406276880551013092746735413);
+        __maxBorrow_correctReturnValue(0);
+        __maxWithdraw_correctMax(0);
     }
 }